Added tests for creating the StandardGradeBook and RankedGradeBook classes, updated instructions to reflect changes
